Use uin32_t for lengths

+++ include/qmedia/neo_media_client.hh
@@ -42,14 +42,14 @@ extern "C"
     void  MediaClient_sendAudio(void *instance,
                                 uint64_t media_stream_id,
                                 const char *buffer,
-                                uint16_t length,
+                                uint32_t length,
                                 uint64_t timestamp);
 
     void
     MediaClient_sendVideoFrame(void *instance,
                                uint64_t streamId,
                                const char *buffer,
-                               uint16_t length,
+                               uint32_t length,
                                uint64_t timestamp,
                                bool flag);
